requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
imagehash>=4.3.1
easyocr>=1.7.0
scikit-image>=0.21.0

